requirements.txt

langchain>=0.2.10
langchain-community>=0.2.10
langgraph>=0.2.0
langgraph-checkpoint>=1.0.0
pymilvus>=2.4.0
redis>=5.0.0
rank-bm25>=0.2.2
pdfplumber>=0.11
pdfminer.six>=20231228
pdf2image>=1.17.0
pytesseract>=0.3.10
Pillow>=10.3.0
cerbos>=0.7.0
reportlab>=4.0.7
